refactor(DC-704): migrate Outage startup test onto StartupValidationTestBase
